created new directory for voucher card and moved consts into types file and imported them
