+++ packages/f-content-cards/src/components/cardTemplates/index.js
@@ -3,7 +3,7 @@ import PromotionCard from './PromotionCard.vue';
 import PostOrderCard from './PostOrderCard.vue';
 import SkeletonLoader from './SkeletonLoader.vue';
 import TermsAndConditionsCard from './TermsAndConditionsCard.vue';
-import VoucherCard from './VoucherCard.vue';
+import VoucherCard from './voucherCard/VoucherCard.vue';
 import { HomePromotionCard1, HomePromotionCard2 } from './homePromotionCard';
 
 export default {
