+++ packages/f-content-cards/src/components/cardTemplates/voucherCard/service.types.js
@@ -1,0 +1,3 @@
+export const COPY_STATE_AVAILABLE = 'available';
+export const COPY_STATE_COOLDOWN = 'cooldown';
+export const COPY_STATE_TRANSITIONOUT = 'transitionout';
